requirements.txt
streamlit
pandas
duckdb
openpyxl
plotly
pyarrow
psycopg2-binary